// src/types.ts
export enum Language {
    EN = 'en',
    PT = 'pt',
    FR = 'fr',
    ES = 'es',
    ZH = 'zh',
    JA = 'ja',
    EN_GB = 'en-GB',
}

export enum Mode {
    FULLSTACK = 'fullstack',
    FRONTEND = 'frontend',
    BACKEND = 'backend',
    DEVOPS = 'devops',
    DATABASE = 'database',
    AI = 'ai',
}